# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_smart-invest-solutions_backend_internal_domain.CreateGeneralInsuranceDTO:
    properties:
      advisor_contact:
        type: string
      advisor_name:
        type: string
      company_name:
        type: string
      date_of_expiry:
        type: string
      policy_no:
        type: string
      vehicle_no:
        type: string
    required:
    - company_name
    - date_of_expiry
    - policy_no
    - vehicle_no
    type: object
...
  github_com_smart-invest-solutions_backend_internal_domain.GeneralInsurance:
    properties:
      advisor_contact:
        type: string
      advisor_name:
        type: string
      company_name:
        type: string
      created_at:
        type: string
      date_of_expiry:
        description: 'Format: YYYY-MM-DD'
        type: string
      id:
        type: string
      policy_no:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      vehicle_no:
        type: string
    required:
    - company_name
    - date_of_expiry
    - policy_no
    - vehicle_no
    type: object
  github_com_smart-invest-solutions_backend_internal_domain.GeneralInsuranceListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.GeneralInsurance'
        type: array
      total:
        type: integer
    type: object
  github_com_smart-invest-solutions_backend_internal_domain.GeneralInsuranceWithCustomer:
    properties:
      advisor_contact:
        type: string
      advisor_name:
        type: string
      agency_id:
        description: |-
          AgencyID is the owning customer's Agency ID — surfaced so a Super Admin can see which admin's
          agency each policy belongs to. Empty for unassigned clients.
        type: string
      company_name:
        type: string
      contact_no:
        type: string
      created_at:
        type: string
      customer_name:
        type: string
      date_of_expiry:
        type: string
      id:
        type: string
      policy_no:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      vehicle_no:
        type: string
    type: object
...
  github_com_smart-invest-solutions_backend_internal_domain.UpdateGeneralInsuranceDTO:
    properties:
      advisor_contact:
        type: string
      advisor_name:
        type: string
      company_name:
        type: string
      date_of_expiry:
        type: string
      policy_no:
        type: string
      vehicle_no:
        type: string
    type: object
...
  github_com_smart-invest-solutions_backend_pkg_response.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_smart-invest-solutions_backend_pkg_response.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      success:
        type: boolean
      total:
        type: integer
      total_pages:
        type: integer
    type: object
info:
  contact:
    email: support@ulmind.com
    name: ULMIND Team
  description: |-
    Backend API for Smart Invest Solutions — a LIC Policy & Investment Management Platform.
    This API provides user registration, authentication with JWT, role-based access control, and more.
  title: Smart Invest Solutions API
  version: "1.0"
paths:
...
  /general-insurances:
    get:
      consumes:
      - application/json
      description: Retrieves list of all general/vehicle insurance policies for the
        authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: General insurance policies retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.GeneralInsuranceListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all my General Insurance policies
      tags:
      - General Insurance
    post:
      consumes:
      - application/json
      description: Creates a new general/vehicle insurance policy linked to the authenticated
        user account.
      parameters:
      - description: General Insurance Policy Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.CreateGeneralInsuranceDTO'
      produces:
      - application/json
      responses:
        "201":
          description: General Insurance policy added successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.GeneralInsurance'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Add General Insurance policy
      tags:
      - General Insurance
  /general-insurances/{id}:
    delete:
      consumes:
      - application/json
      description: Removes a general insurance policy from the authenticated user
        account.
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Policy deleted successfully
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete General Insurance policy
      tags:
      - General Insurance
    get:
      consumes:
      - application/json
      description: Retrieves detailed info for a single general insurance policy.
      parameters:
      - description: General Insurance ID (MongoDB ObjectID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Policy retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.GeneralInsurance'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "404":
          description: Policy not found
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get General Insurance policy by ID
      tags:
      - General Insurance
    put:
      consumes:
      - application/json
      description: Updates vehicle no, policy no, date of expiry, company name, or
        advisor info.
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.UpdateGeneralInsuranceDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Policy updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.GeneralInsurance'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "404":
          description: Policy not found
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update / Edit General Insurance policy
      tags:
      - General Insurance
  /general-insurances/all:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated master list of every general/vehicle insurance
        policy — Customer Name, Contact No, Agency ID, Vehicle No, Policy No, Date
        of Expiry, Company Name. A super_admin receives every client's policies; a
        plain admin receives one scoped to clients under their own Agency ID.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Policies retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.GeneralInsuranceWithCustomer'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "403":
          description: Forbidden — Admin role required
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all clients' General Insurance policies (Admin Only)
      tags:
      - General Insurance (Admin)
  /general-insurances/user/{userId}:
    get:
      consumes:
      - application/json
      description: Retrieves all general insurance policies associated with a specific
        user ID for Admin/Advisor view.
      parameters:
      - description: Target Client User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Policies retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.GeneralInsuranceListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "403":
          description: Forbidden — Admin role required
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user's General Insurance policies (Admin Only)
      tags:
      - General Insurance (Admin)
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your Bearer token in the format: **Bearer &lt;token&gt;**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"